+++ client/src/components/shared/Nav/Nav.jsx
@@ -4,6 +4,8 @@ function Nav(props) {
   return (
     <div className="nav-div">
       <p>Here is our Navo</p>
+      <Link to="/">Home</Link>
+      <Link to="/add">Add</Link>
     </div>
   );
 }
